Add Photon ID section to check results

+++ quantile_regression_chain/quantileRegression_chain.py
@@ -640,9 +640,9 @@ class quantileRegression_chain(object):
         """
 
         if key=='mc':
-            stride = self.MC.index.size / n_jobs
+            stride = int(self.MC.index.size / n_jobs)
         elif key=='data':
-            stride = self.data.index.size / n_jobs
+            stride = int(self.data.index.size / n_jobs)
         logger.info("Computing %s, correcting %s, stride %s" % (name,correctedVariables,stride) )
         if key == 'mc':
             with parallel_backend(self.backend):
